+++ src/main.cpp
@@ -14,6 +14,7 @@
 #include "xcore/sound.h"
 #include "xcore/pacing.h"
 #include "xcore/autostart.h"
+#include "xcore/vfat_scan.h"
 #include "xgui/xgui.h"
 #include "libxpeccy/spectrum.h"
 #include "libxpeccy/cpu/Z80/z80.h"
@@ -54,7 +55,7 @@ void help() {
 	printf("--bp NAME\t\tset fetch brakepoint to label NAME (see -l key)\n");
 	printf("--brk FILE\t\tload breakpoints list (*.xbrk)\n");
 	printf("--disk X\t\tselect drive to loading file (0..3 | a..d | A..D)\n");
-	printf("--sdcard FILE\t\tset SD card image\n");
+	printf("--sdcard PATH\t\tset SD card image, or a folder served as one\n");
 	printf("--style\t\t\tMacOSX only: use native qt style, else 'fusion' will be forced\n");
 	printf("--xmap FILE\t\tLoad *.xmap file\n");
 	printf("--confdir DIR\t\tChange config directory\n");
@@ -367,7 +368,7 @@ int main(int ac,char** av) {
 				conf.vid.fullScreen = atoi(av[i]) ? 1 : 0;
 				i++;
 			} else if (!strcmp(parg, "--sdcard")) {
-				sdcSetImage(conf.prof.cur->zx->sdc, av[i]);
+				sdc_mount(conf.prof.cur->zx->sdc, QString::fromLocal8Bit(av[i]));
 				i++;
 			} else if (!strcmp(parg, "--disk")) {
 				parg = av[i];